Moved integration tests into integration test suite.
- suppressed test dep warnings.
